docs: make demos tsx
