+++ package.json
@@ -39,7 +39,7 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@superset-ui/build-config": "^0.0.4",
-    "@superset-ui/chart": "0.10.2",
+    "@superset-ui/chart": "^0.10.8",
     "@superset-ui/color": "^0.10.0",
     "@superset-ui/connection": "^0.10.0",
     "@superset-ui/core": "^0.10.0",
@@ -47,6 +47,7 @@
     "@superset-ui/number-format": "^0.10.0",
     "@superset-ui/time-format": "^0.10.0",
     "@superset-ui/translation": "^0.10.0",
+    "@types/react": "^16.8.8",
     "csstype": "^2.6.3",
     "fast-glob": "^2.2.6",
     "fs-extra": "^7.0.1",
