+++ packages/superset-ui-plugins-demo/package.json
@@ -35,10 +35,11 @@
     "@storybook/addon-knobs": "^4.0.2",
     "@storybook/addon-options": "^4.0.3",
     "@storybook/react": "^4.1.11",
-    "@superset-ui/chart": "0.10.2",
+    "@superset-ui/chart": "^0.10.8",
     "@superset-ui/color": "^0.10.1",
     "@superset-ui/time-format": "^0.10.1",
     "@superset-ui/translation": "^0.10.0",
+    "@types/react": "^16.8.8",
     "babel-loader": "^8.0.4",
     "bootstrap": "^3.3.6",
     "cache-loader": "^1.2.2",
