+++ packages/superset-ui-plugins-demo/storybook/stories/preset-chart-xy/Line/constants.ts
@@ -1,0 +1,2 @@
+export const LINE_PLUGIN_TYPE = 'v2-line';
+export const LINE_PLUGIN_LEGACY_TYPE = 'v2-line/legacy';
