+++ packages/superset-ui-plugins-demo/storybook/stories/preset-chart-xy/Line/index.js
@@ -4,9 +4,10 @@ import BasicStories from './stories/basic';
 import LegacyStories from './stories/legacy';
 import MissingStories from './stories/missing';
 import TimeShiftStories from './stories/timeShift';
+import { LINE_PLUGIN_TYPE, LINE_PLUGIN_LEGACY_TYPE } from './constants';
 
-new LegacyLineChartPlugin().configure({ key: 'v2-line/legacy' }).register();
-new LineChartPlugin().configure({ key: 'v2-line' }).register();
+new LegacyLineChartPlugin().configure({ key: LINE_PLUGIN_LEGACY_TYPE }).register();
+new LineChartPlugin().configure({ key: LINE_PLUGIN_TYPE }).register();
 
 export default {
   examples: [...BasicStories, ...MissingStories, ...TimeShiftStories, ...LegacyStories],
