--- packages/superset-ui-plugins-demo/storybook/stories/preset-chart-xy/Line/stories/legacy.jsx
@@ -1,73 +1,0 @@
-/* eslint-disable no-magic-numbers */
-import React from 'react';
-import { SuperChart } from '@superset-ui/chart';
-import data from '../data/legacyData';
-
-export default [
-  {
-    renderStory: () => [
-      <SuperChart
-        key="line1"
-        chartType="v2-line/legacy"
-        chartProps={{
-          datasource: { verboseMap: {} },
-          formData: {
-            bottomMargin: 'auto',
-            colorScheme: 'd3Category10',
-            leftMargin: 'auto',
-            lineInterpolation: 'linear',
-            richTooltip: true,
-            showBrush: 'auto',
-            showLegend: true,
-            showMarkers: false,
-            vizType: 'line',
-            xAxisFormat: '%Y',
-            xAxisLabel: '',
-            xAxisShowminmax: false,
-            xTicksLayout: 'auto',
-            yAxisBounds: [null, null],
-            yAxisFormat: '',
-            yAxisLabel: '',
-            yAxisShowminmax: false,
-            yLogScale: false,
-          },
-          height: 400,
-          payload: { data },
-          width: 400,
-        }}
-      />,
-      <SuperChart
-        key="line2"
-        chartType="v2-line/legacy"
-        chartProps={{
-          datasource: { verboseMap: {} },
-          formData: {
-            bottomMargin: 'auto',
-            colorScheme: 'd3Category10',
-            leftMargin: 'auto',
-            lineInterpolation: 'linear',
-            richTooltip: true,
-            showBrush: 'auto',
-            showLegend: true,
-            showMarkers: false,
-            vizType: 'line',
-            xAxisFormat: '%Y-%m',
-            xAxisLabel: '',
-            xAxisShowminmax: false,
-            xTicksLayout: 'auto',
-            yAxisBounds: [null, null],
-            yAxisFormat: '',
-            yAxisLabel: '',
-            yAxisShowminmax: false,
-            yLogScale: false,
-          },
-          height: 400,
-          payload: { data },
-          width: 800,
-        }}
-      />,
-    ],
-    storyName: 'Use Legacy API shim',
-    storyPath: 'preset-chart-xy|LineChartPlugin',
-  },
-];
